requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.5.0
dill==0.3.8
et_xmlfile==2.0.0
faiss-cpu==1.5.3
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.12.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.2.1
numpy<2.0
openpyxl==3.1.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pyarrow==19.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
